Refactor Rust libraries
Follow up for #5645 and  https://github.com/edgedb/edgedb-cli/pull/1080.

- removes "python" from default features of edgeql-parser,
- renames edgeql-rust to edgeql-parser-python,
- moves edgeql-parser-* libraries under edgeql-parser dir,
