+++ Cargo.toml
@@ -1,8 +1,8 @@
 [workspace]
 members = [
     "edb/edgeql-parser",
-    "edb/edgeql-parser-derive",
-    "edb/edgeql-rust",
+    "edb/edgeql-parser/edgeql-parser-derive",
+    "edb/edgeql-parser/edgeql-parser-python",
     "edb/graphql-rewrite",
 ]
