+++ edb/edgeql-parser/Cargo.toml
@@ -18,12 +18,12 @@ wasm-bindgen = { version = "0.2", features = [
 serde = { version = "1.0.106", features = ["derive"], optional = true }
 thiserror = "1.0.23"
 unicode-width = "0.1.8"
-edgeql-parser-derive = { path = "../edgeql-parser-derive", optional = true }
+edgeql-parser-derive = { path = "edgeql-parser-derive", optional = true }
 cpython = { version = "0.7.0", optional = true }
 indexmap = "1.9.3"
 
 [features]
-default = ["python"]
+default = []
 wasm-lexer = ["wasm-bindgen", "serde"]
 python = ["cpython", "edgeql-parser-derive"]
