+++ edb/edgeql-parser/edgeql-parser-python/src/keywords.rs
