+++ edb/edgeql-parser/edgeql-parser-python/tests/normalize.rs
